# Reimplement with `os.scandir` for 2.7x speedup