This commit reimplements git-find-repos to use `os.scandir` to list directory entries, instead of `pathlib.Path.iterdir`.

Using `os.scandir` significantly increases performance because the `os.DirEntry` objects returned expose file attribute information, when the operating system provides it, without additional system calls. For
example, calling the `is_dir` method does not require an additional system call to be made except in the case of symlinks.

In benchmarking this change in the home directory of my development machine, which contains around 80 Git repositories at various different directory depths, I observe a speedup of 2.7x:

    % hyperfine './git-find-repos-orig $HOME' './git-find-repos-new $HOME'
    Benchmark 1: ./git-find-repos-orig $HOME
      Time (mean ± σ):      2.286 s ±  0.032 s    [User: 1.329 s, System: 0.920 s]
      Range (min … max):    2.256 s …  2.355 s    10 runs

    Benchmark 2: ./git-find-repos-new $HOME
      Time (mean ± σ):     852.1 ms ±  15.0 ms    [User: 259.0 ms, System: 537.1 ms]
      Range (min … max):   826.9 ms … 875.4 ms    10 runs

    Summary
      './git-find-repos-new $HOME' ran
        2.68 ± 0.06 times faster than './git-find-repos-orig $HOME'