@acroz 2022-01-24
Released as [2.1.0](https://pypi.org/project/git-find-repos/2.1.0/)